+++ OTA_CLI/Programmer/UpdaterInterface.cs
@@ -2,6 +2,6 @@ namespace OTA_CLI;
 
 public interface UpdaterInterface
 {
-    public bool Update(int targetDeviceId);
+    public bool Update(byte targetDeviceId);
     public bool Verify();
 }
